requirements.txt
# Exchange & Market Data
ccxt==4.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Technical Analysis
ta==0.11.0

# Telegram Bot
python-telegram-bot==20.7

# Database
sqlalchemy==2.0.23

# Environment Variables
python-dotenv==1.0.0

# Scheduling
APScheduler==3.10.4

# Utilities
requests==2.31.0